Cleaned up agents and race control
