+++ Assets/Scenes/RacePhysics.unity
@@ -127,10 +127,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 174661994}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2392fc7134099d94b97394ccc7e52d03, type: 3}
+  m_Script: {fileID: 11500000, guid: ab7ea4c1831cfad409758b3f5082b096, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  raceControl: {fileID: 1317049839}
+  raceControl: {fileID: 1317049841}
 --- !u!65 &174661999
 BoxCollider:
   m_ObjectHideFlags: 0
@@ -139,8 +139,8 @@ BoxCollider:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 174661994}
   m_Material: {fileID: 0}
-  m_IsTrigger: 0
-  m_Enabled: 0
+  m_IsTrigger: 1
+  m_Enabled: 1
   serializedVersion: 2
   m_Size: {x: 5.253531, y: 2.904412, z: 4.00251}
   m_Center: {x: -1.1745596, y: 0, z: -1.176347}
@@ -150,50 +150,6 @@ Transform:
     type: 3}
   m_PrefabInstance: {fileID: 1566082218}
   m_PrefabAsset: {fileID: 0}
---- !u!1 &194199979
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 194199980}
-  - component: {fileID: 194199981}
-  m_Layer: 0
-  m_Name: Goal
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &194199980
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 194199979}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 241.6, y: -18.4, z: -29.1}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_RootOrder: 7
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!65 &194199981
-BoxCollider:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 194199979}
-  m_Material: {fileID: 0}
-  m_IsTrigger: 0
-  m_Enabled: 1
-  serializedVersion: 2
-  m_Size: {x: 20, y: 10, z: 40}
-  m_Center: {x: 0, y: 0, z: 0}
 --- !u!1 &213455487
 GameObject:
   m_ObjectHideFlags: 0
@@ -1047,12 +1003,6 @@ Transform:
   m_Father: {fileID: 1604142274}
   m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!223 &826374407 stripped
-Canvas:
-  m_CorrespondingSourceObject: {fileID: 6324481468387304104, guid: e27a279b1112bd8498aff4cc38b81d91,
-    type: 3}
-  m_PrefabInstance: {fileID: 6324481469716283836}
-  m_PrefabAsset: {fileID: 0}
 --- !u!1 &908072305
 GameObject:
   m_ObjectHideFlags: 0
@@ -1573,7 +1523,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 8
+  m_RootOrder: 7
   m_LocalEulerAnglesHint: {x: 0, y: 90.01, z: 0}
 --- !u!1 &1289374081
 GameObject:
@@ -1675,7 +1625,6 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 1317049840}
-  - component: {fileID: 1317049839}
   - component: {fileID: 1317049841}
   m_Layer: 0
   m_Name: RaceControl
@@ -1684,21 +1633,6 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!114 &1317049839
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1317049838}
-  m_Enabled: 0
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 40a5e863f7274b14caca248422d357bd, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  uiCanvas: {fileID: 826374407}
-  leaderboard: {fileID: 818343476138966343, guid: 9cb1b1238c1b08c49b16ecc44ee639b1,
-    type: 3}
 --- !u!4 &1317049840
 Transform:
   m_ObjectHideFlags: 0
@@ -1725,9 +1659,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 5a8ffcf7126feca4e9d5bef41dc1d436, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  goal: {fileID: 194199980}
+  goal: {fileID: 174662000}
   start: {fileID: 1204752115}
-  numAi: 1
+  numAI: 3
   gliderPrefab: {fileID: 7791143779684347089, guid: c555d35f1806120428c0fe776d93691c,
     type: 3}
   indicator: {fileID: 971863099}
