--- Assets/Scripts/Racing/Agents/AIPhysicsAgent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 23c9a6bda78967942a61c4a3fddb5512
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
