--- Assets/Scripts/Racing/Agents/CheaterAgent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3175299f87f784e4a97a8471af0295fe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
