--- Assets/Scripts/Racing/Agents/PhysicsAgent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9794cbf9191e9494da3f2681b624e665
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
