--- Assets/Scripts/Racing/Agents/PlayerPhysicsAgent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 86bc20ad1cbae9b45997d966a8b05d9c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
