--- Assets/Scripts/Racing/GliderV2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f8b63b8c74562bb489fa3e3ce13bcd3f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
