--- Assets/Scripts/Racing/RaceControl/RaceBeginState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dd5e226bc2943d3478ebdd7e3a0426ef
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
