--- Assets/Scripts/Racing/RaceControl/RaceEndState.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace Racing
-{
-    public class RaceEndState : RaceState
-    {
-        public RaceEndState(RaceStateInput raceInput) : base(raceInput) { }
-        override public void updateState()
-        {
-            Debug.Log("RaceEndState");
-        }
-    }
-
-}
