--- Assets/Scripts/Racing/RaceControl/RaceEndState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f265cfc4ed6e4e341b5ea1649d50ef33
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
