--- Assets/Scripts/Racing/RaceControl/RaceInProgressState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 826c7d1be1fb8a649bad998e2e0da638
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
