--- Assets/Scripts/Racing/RaceControl/RacePhysicsControl.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5a8ffcf7126feca4e9d5bef41dc1d436
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
