--- Assets/Scripts/Racing/RaceControl/RaceStateInput.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 67b8cd80ac43cbd4f9b4a57791713834
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
