--- Assets/Scripts/Racing/RaceEnd.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2392fc7134099d94b97394ccc7e52d03
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
