+++ ProjectSettings/EditorBuildSettings.asset
@@ -14,4 +14,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/Race.unity
     guid: 804f8eabc4ee0574985e089b223076d5
+  - enabled: 1
+    path: Assets/Scenes/Raise.unity
+    guid: 30898962d9c7885419a5f2ef96c4b64f
   m_configObjects: {}
